Create initial project layout and basic configs
This project is in a very barebones state, lacking a defined layout for
the simulation study.

This PR adds a defined environment.yml file, basic README's when
applicable, temporary hidden `.keep` files to include "empty"
directories, and basic configuration for `pre-commit` to ensure that all
users are abiding by a standard style.

The `.keep` files can be removed once other files are included in these
directories.