Initialize test suite for project
We were currently missing any tests and test framework to ensure that
our molecules, etc. will be properly tested/expected to behave as we
anticipate.

This PR sets up the base_test.py class, as well as a set of tests for
the included MethaneUA compound. With the hope that other molecules as
they are added will adhere to a similar set of tests.

This also adds the necessary __init__.py files so we can access this
project as python modules.