Include SPC/E water model and forcefield XML
This includes the foyer xml for SPC/E water and tests to ensure that a
water molecule is being properly parametrized.

Tests are included on a single water molecule to ensure that not only
are the unit conversions are what we expect, we use the
foyer.forcefield.get_parameters method to access the forcefield
information programmatically.

This is a template layout for other molecules and forcefields if needed.