mcccs individual project.py (#18)
* Update project/README.md

Added the dashboard file, signac.rc, and the workspace

* Add some job initialization code to init.py

The purpose of this commit is to get some schema ideas down and to get
the initialization code in place.

* Add dashboard file

* Add signac.rc and workspace

* Add signac flow project

* Add replica to schema

* Update project/init.py

* Update project/init.py

* Add skeleton code for hoomd simulations

* Update project/src/project.py

* Add directory for engine subprojects

* Include SPC/E water model and forcefield XML

This includes the foyer xml for SPC/E water and tests to ensure that a
water molecule is being properly parametrized.

Tests are included on a single water molecule to ensure that not only
are the unit conversions are what we expect, we use the
foyer.forcefield.get_parameters method to access the forcefield
information programmatically.

This is a template layout for other molecules and forcefields if needed.

* add restart file maker for mcccs

* Update project.py

* Remove top-level project.py file

We decided to break the project down into separate subprojects, one for
each of the different simulation engines. This top-level project.py file
is therefore no longer needed.

* Apply pre-commit changes to README

* Apply pre-commit changes to dashboary.py

* Apply pre-commit changes to hoomd/project.py

* Apply pre-commit changes to each project.py

* Remove engines/project.py

Each subproject has its own project.py file in its respective directory,
and therefore this file is not needed.

* update project.py for mcccs simulations

* update doctring for __init__.py

* add copy function for topmon

* Add some job initialization code to init.py

The purpose of this commit is to get some schema ideas down and to get
the initialization code in place.

* Add dashboard file

* Add signac flow project

* Add replica to schema

* Update project/init.py

* Update project/init.py

* Add skeleton code for hoomd simulations

* Update project/src/project.py

* Add directory for engine subprojects

* add restart file maker for mcccs

* Update project.py

* Remove top-level project.py file

We decided to break the project down into separate subprojects, one for
each of the different simulation engines. This top-level project.py file
is therefore no longer needed.

* Apply pre-commit changes to dashboary.py

* Apply pre-commit changes to hoomd/project.py

* Apply pre-commit changes to each project.py

* Remove engines/project.py

Each subproject has its own project.py file in its respective directory,
and therefore this file is not needed.

* update project.py for mcccs simulations

* update doctring for __init__.py

* add copy function for topmon

* start fort4 for pentane

Co-authored-by: Tim Moore <mtimc@umich.edu>
Co-authored-by: Justin Gilmer <justin.b.gilmer@gmail.com>
Co-authored-by: Justin Gilmer <justin.b.gilmer@vanderbilt.edu>