Add more clever filtering to init.py
Currently, we use a naive combinatoric product to generate all the
possible combinations of our statepoints.

However, this method, although fast, does not include any chemical
intuition or other filtering necessary for this project.
For example, all simulation engines have a statepoint associated with a
GEMC-NVT simulation ensemble, something not possible with the molecular
dynamics engines included in the study, only the monte-carlo engines.