Maint/initproject modify (#39)
* Update project/README.md

Added the dashboard file, signac.rc, and the workspace

* Add some job initialization code to init.py

The purpose of this commit is to get some schema ideas down and to get
the initialization code in place.

* Add dashboard file

* Add signac.rc and workspace

* Add signac flow project

* Add replica to schema

* Update project/init.py

* Update project/init.py

* Add skeleton code for hoomd simulations

* Update project/src/project.py

* Add directory for engine subprojects

* Update project.py

* Remove top-level project.py file

We decided to break the project down into separate subprojects, one for
each of the different simulation engines. This top-level project.py file
is therefore no longer needed.

* Apply pre-commit changes to README

* Apply pre-commit changes to dashboary.py

* Apply pre-commit changes to hoomd/project.py

* Apply pre-commit changes to each project.py

* Remove engines/project.py

Each subproject has its own project.py file in its respective directory,
and therefore this file is not needed.

* Add more clever filtering to init.py

Currently, we use a naive combinatoric product to generate all the
possible combinations of our statepoints.

However, this method, although fast, does not include any chemical
intuition or other filtering necessary for this project.
For example, all simulation engines have a statepoint associated with a
GEMC-NVT simulation ensemble, something not possible with the molecular
dynamics engines included in the study, only the monte-carlo engines.

* Create helper variables for md and mc engines

To make the logic a bit simpler when filtering the non-useful
statepoints, some helper lists were created to make the logic a bit
cleaner. Thanks to @jennyfothergill for the suggestion!

* pre-commit fixes

Co-authored-by: Tim Moore <mtimc@umich.edu>