Overhaul statepoint generation, addl. info added (#49)
* Overhaul statepoint generation, addl. info added

Previously, we were using a standard combinatoric product to generate
all the possible statepoints for the project.

This would work correctly, if each parameter we are changing applies to
every system. However, this is not the case. This overhaul attempts to
provide a bit more of an extensible setup for each molecule, as well as
being a bit more tunable based on the specific statepoint information.

* Include r_cut, cutoff style, vap and liq densities

* Round unit converted values to 3 decimal places