Create plots with t0 as a vertical line
To make quick inspection of trajectories and their progress towards
equilibration, being able to quickly generate plots over a data space to
inspect with signac dashboard will make our lives much easier.

This is not feature complete, features like matplotlib plotting kwargs
and other options will be exposed to the user, which will be implemented
in the analysis module.