Include Neff as another check in is_equilibrated
Previously, the only metric in `is_equilibrated` to determine if a set
of data was sufficiently equilibrated was only with a fraction of the
total dataset.

This creates issues when dealing with very small data sets since we also
require a sufficient amount of effectively uncorrelated samples to do
proper statistical treatments. Now, there are two values that need to be
true before we classify a dataset as equilibrated:

1. A fraction of the dataset must be considered effectively decorrelated
(the starting index of the location in the data that is the first
uncorrelated sample)
2. A minimum amount of uncorrelated samples needs to be available based
   on the starting index and the statistical inefficiency of the data.

This PR adds number 2, since number 1 was included in the first
iteration of this method (it has been named from `threshold` to
`threshold_fraction`.
A new function argument, `threshold_neff` has been added, with a default
value of 100.

Now, for `is_equilibrated` to return `True`, it must satisfy both of
these threshold values to be considered equilibrated.