Write out subsampled data points to job.data
We currently have the ability to write out the indices of the subsampled
data that is effectively decorrelated.

This stores the values of the subsampled property in the `job.data`
object (see https://docs.signac.io/en/latest/jobs.html#job-data-storage
) for more information.

The values of the property are located at:
`job.data["subsamples/PROPERTY_OF_INTEREST"]`. For example:
`job.data["subsamples/density"]`. This provides a numpy-like access to
tabular data and scalar values.