Update sample_job, write sampled vals to job.data
Previously we would attempt to store a range object to the job document,
this might not work since it cannot be directly converted to a json
object.

This snippet below will fail
```py3
import json
my_dict = {"foo": range(0,10,2)}
json.dumps(my_dict) # error
```

This has been changed to a dictionary where the `start`, `stop`, and
`step` values of the range constructor are key/value pairs in a
dictionary.
This will work with `json` dumping and we can re-generate the range
object when needed.

An additional method has been added to take this information and then
write out the subsamples to the job.data object.