Write out subsampled data points to job.data (#95)
* Write out subsampled data points to job.data

We currently have the ability to write out the indices of the subsampled
data that is effectively decorrelated.

This stores the values of the subsampled property in the `job.data`
object (see https://docs.signac.io/en/latest/jobs.html#job-data-storage
) for more information.

The values of the property are located at:
`job.data["subsamples/PROPERTY_OF_INTEREST"]`. For example:
`job.data["subsamples/density"]`. This provides a numpy-like access to
tabular data and scalar values.

* Fix typo

* Update sample_job, write sampled vals to job.data

Previously we would attempt to store a range object to the job document,
this might not work since it cannot be directly converted to a json
object.

This snippet below will fail
```py3
import json
my_dict = {"foo": range(0,10,2)}
json.dumps(my_dict) # error
```

This has been changed to a dictionary where the `start`, `stop`, and
`step` values of the range constructor are key/value pairs in a
dictionary.
This will work with `json` dumping and we can re-generate the range
object when needed.

An additional method has been added to take this information and then
write out the subsamples to the job.data object.

* Include unit tests and install h5py for job.data use