Include analysis routines to write sampled data
Currently, we have utilities to subsample the properties of interest for
statistical treatment, however, we have not implemented this in our
analysis signac project.

This is the initial addition of these routines for both NPT and NVT
production data.

Additional unit tests are needed, as well as further normalization of
the data between each engine (specifically column names), but since we
have the data for methane manually calculated, this should be a great
way to test this.