Create subproject to explore LRC and shifting
Our initial results from the TraPPE United-Atom methane simulations
(`MethaneUA`) had discrepancies on the order of a few percentage points
for the density values when comparing MD vs MC simulations.

This appears to be a factor due to long range corrections (tail
corrections) and how the potential is truncated at the cutoff (abruptly
change to 0 once you reach r_cut) or shifted (a constant term is applied
to make the value of the potential at r_cut 0).

A subproject has been created to further explore this for MethaneUA for
all engines that support LRC and shifted potentials. This PR sets up the
groundwork for this.