Remove statepoints with LRC and shifted potentials (#116)
It does not make sense to run a simulation with both long range
corrections and shifted potentials at the same time, in fact, LAMMPS
forbids this entirely.

LRC's are meant to be used when there are discontinuities at the cutoff,
and shifted potentials are meant to transition the potential to 0 at the
cutoff. These attempt to mitigate similar issues.