Change write_subsampled_values to return an array
Previously, the write_subsampled_values method operated on a job and
then wrote to the job.data store.

This is not very generic and more difficult to apply to other cases
where the subsampled values might be useful to have passed around as a
numpy array. This is now the default case.