Include analysis routines to write sampled data (#105)
* Include analysis routines to write sampled data

Currently, we have utilities to subsample the properties of interest for
statistical treatment, however, we have not implemented this in our
analysis signac project.

This is the initial addition of these routines for both NPT and NVT
production data.

Additional unit tests are needed, as well as further normalization of
the data between each engine (specifically column names), but since we
have the data for methane manually calculated, this should be a great
way to test this.

* Separate multiple steps into functions: 'sample_job', 'write_sampled_data', 'average_sampled_data'

* Add safer method to access nested dicts

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add cassandra2gsd function. (#106)

* Add cassandra2gsd function.

* Evaluate scalar logical instead of list in if statement.

* Replace all with any in if condition.

* Open gsdfile directly from path object

Co-authored-by: Jenny <39961845+jennyfothergill@users.noreply.github.com>

* Fix types definition

Co-authored-by: Jenny <39961845+jennyfothergill@users.noreply.github.com>

* Reformat path object and file object names in snake_case.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fill out docstrings.

* Import hoomd.

* Add optional arguments allowing length unit selection.

* pre-commit fix

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* pre-commit fix

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Co-authored-by: Jenny <39961845+jennyfothergill@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Justin Gilmer <justin.b.gilmer@vanderbilt.edu>

* Create subproject to explore LRC and shifting (#110)

* Create subproject to explore LRC and shifting

Our initial results from the TraPPE United-Atom methane simulations
(`MethaneUA`) had discrepancies on the order of a few percentage points
for the density values when comparing MD vs MC simulations.

This appears to be a factor due to long range corrections (tail
corrections) and how the potential is truncated at the cutoff (abruptly
change to 0 once you reach r_cut) or shifted (a constant term is applied
to make the value of the potential at r_cut 0).

A subproject has been created to further explore this for MethaneUA for
all engines that support LRC and shifted potentials. This PR sets up the
groundwork for this.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Include SPCE water and OPLSAA ethanol statepoints

* add structure for engine project.py if needed

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>

* Remove extraneous temperatures and pressure from subproject-init.py (280K, 320K) (#114)

* Update Lammps project for nvt and npt simulations (#113)

* Modify lammps project.py to output data from production trajectory

	* Write out and check for log.txt
	* Convert units to those used in the rest of the project
	* Write out a box.gro for import to mdtraj
	* Use analysis tools to calculate rdf and diffusion coefficient from trajectory.gsd
	* Use sampler.py to identify uncorrelated data from thermo output

* Add 5000 steps to in.equilibration, add timestep variable to input scripts based on molecule

* Update lammps project.py to include run nvt and npt production steps, instead of just one npt production

	* Create in.production-npt, in.production-nvt
	* Fix bug with writing out log-npt.txt to pass pe in ke column
	* Create project.py operation for running nvt step
	* Create gsd and txt files from nvt step

* Delete old in.production file

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* copy over files from main project src for lammps-VU simulations and make available in subproject directoy lrc_shit_subproject

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Remove repeated defintiion for unit conversions in lammps=VU project.py

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>

* Remove statepoints with LRC and shifted potentials (#116)

It does not make sense to run a simulation with both long range
corrections and shifted potentials at the same time, in fact, LAMMPS
forbids this entirely.

LRC's are meant to be used when there are discontinuities at the cutoff,
and shifted potentials are meant to transition the potential to 0 at the
cutoff. These attempt to mitigate similar issues.

* Simplify the property sampling process

* change write subsampled values to return a numpy array

* Remove older method to sample and average data

* Change write_subsampled_values to return an array

Previously, the write_subsampled_values method operated on a job and
then wrote to the job.data store.

This is not very generic and more difficult to apply to other cases
where the subsampled values might be useful to have passed around as a
numpy array. This is now the default case.

* Include more descriptive labels and more modular functions

* Include sampling based on max t0 value

* Finalize per job statistics.

* Include statistical treatments for all properties of interest, skeleton of aggregate data included

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Ryan Smith <rsmith56@nd.edu>
Co-authored-by: Jenny <39961845+jennyfothergill@users.noreply.github.com>
Co-authored-by: CalCraven <54594941+CalCraven@users.noreply.github.com>