Update benzeneUA ff and utils to TraPPE parameters
Previously, we were using a benzene force field that we meant to be ran
as a flexible model, this led to inconsistent simulations between MC and
MD.

In addition, the flexible models had a very long equilibration time.

This switches the model to the standard TraPPE-UA representation of
benzene, which is treated as a rigid body.